requirements.txt
redis
greenlet>=1.1.0
gevent>=21.1.0
requests
urllib3
pycrypto
Pillow
flask>=2.0.0
flask-script>=2.0.5
flask_bootstrap
PyMySQL>=0.8
flask_sqlalchemy
flask-debugtoolbar
Flask-WTF
Flask-Bcrypt
Flask-Security-too==4.0.1
Flask-Restful>=0.3.9
flask-babelex==0.9.4
flask-babel==2.0.0
flask_migrate==2.7.0
gunicorn
WTForms-Components
werkzeug
waitress
flasgger
xlwt
webassets
sympy
flask-bootstrap
wtforms
# pymongo==3.5.1
# Flask-MongoEngine>=0.9.5
# flask-login==0.3.2
# flask-uploads
# Flask-Session
# Celery
# Flask-Celery-Helper
# Flask-Cache
